chore: 주석 정리

+++ src/main/java/OneQ/OnSurvey/domain/question/service/QuestionCommandService.java
@@ -36,7 +36,7 @@ public class QuestionCommandService implements QuestionCommand {
 
     @Override
     public QuestionUpsertDto upsertQuestionList(QuestionUpsertDto upsertDto) {
-        log.info("[QUESTION:COMMAND:upsertQuestionList] 문항 UPSERT - surveyId: {}, upsertInfoList: {}", upsertDto.getSurveyId(), upsertDto.getUpsertInfoList().toString());
+        log.info("[QUESTION:COMMAND:upsertQuestionList] 문항 UPSERT - surveyId: {}", upsertDto.getSurveyId());
 
         Long surveyId = upsertDto.getSurveyId();
         List<QuestionUpsertDto.UpsertInfo> upsertInfoList = upsertDto.getUpsertInfoList();
@@ -314,7 +314,7 @@ public class QuestionCommandService implements QuestionCommand {
 
     @Override
     public List<OptionUpsertDto> upsertChoiceOptionList(List<OptionUpsertDto> upsertDtoList) {
-        log.info("[QUESTION:COMMAND:upsertChoiceOptionList] 보기 UPSERT - upsertDtoList : {}", upsertDtoList.toString());
+        log.info("[QUESTION:COMMAND:upsertChoiceOptionList] 보기 UPSERT");
 
         List<ChoiceOption> finalList = new ArrayList<>();
 
@@ -345,11 +345,10 @@ public class QuestionCommandService implements QuestionCommand {
                 .map(ChoiceOption::getChoiceOptionId)
                 .filter(optionId -> !updateIdSet.contains(optionId))
                 .collect(Collectors.toSet());
-            log.info("[QUESTION:COMMAND:upsertChoiceOptionList] 삭제되는 문항: {}, 보기 IDs: {}", questionId, deleteIdSet);
-
-            choiceOptionRepository.deleteAll(deleteIdSet);
-
-            log.info("[QUESTION:COMMAND:upsertChoiceOptionList] DELETE 진행");
+            if (!deleteIdSet.isEmpty()) {
+                log.info("[QUESTION:COMMAND:upsertChoiceOptionList] 삭제되는 문항: {}, 보기 IDs: {}", questionId, deleteIdSet);
+                choiceOptionRepository.deleteAll(deleteIdSet);
+            }
 
             // 5. Update 대상 수정
             Map<Long, OptionDto> updateInfoMap = updateInfoList.stream().collect(Collectors.toMap(
